@@ -8,28 +8,30 @@ import { motion } from "framer-motion"
 function Card(props) {
 
     const variants = {
-        visible: i => ({
+        visible: (i) => ({
             opacity: 1,
             y: 0,
             transition: {
-              delay: i * 0.1,
-              duration: 0.5,
-              ease: "easeIn",
-              type: "spring",
-              stiffness: 50
+                delay: i * 0.1,
+                duration: 0.5,
+                ease: "easeIn",
+                type: "spring",
+                stiffness: 50
             },
-          }),
-        hidden: { opacity: 0, y: 200}
-    }
+        }),
+        hidden: { opacity: 0, y: 200 }
+    };
 
     return (
         <a href={props.link}>
-            <motion.div className = "Card four columns"
+            <motion.div className="Card four columns"
                 initial="hidden"
                 animate="visible"
                 custom={props.i}
-                variants={variants}>
-                <img className = "cover" src = {props.cover} alt=""></img>
+                variants={variants}
+            // layout and other props can be added for framer-motion v10+
+            >
+                <img className="cover" src={props.cover} alt="" />
                 <div className="data">
                     <h2>{props.title}</h2>
                     <p>{props.subtitle}</p>
